LayoutTests/platform/mac-wk2/imported/w3c/web-platform-tests/cookiestore/change_eventhandler_for_no_change.https.window-expected.txt


FAIL CookieStore duplicate cookie should not be observed assert_equals: Duplicate cookie is not observed.: changed cookie name expected "alt-cookie" but got "cookie"
FAIL CookieStore duplicate partitioned cookie should not be observed assert_equals: Duplicate cookie is not observed.: changed cookie name expected "alt-cookie" but got "cookie"
